+++ packages/datadog_openfeature_provider/lib/datadog_openfeature_provider.dart
@@ -1,0 +1,9 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0. This product includes software
+// developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2019-Present Datadog, Inc.
+
+/// Datadog provider for the OpenFeature Dart client SDK.
+library;
+
+export 'src/datadog_openfeature_provider.dart' show DatadogOpenFeatureProvider;